Revert "split tests for speedier tests"
This reverts commit 22cffcc80053fb1c3609fcd0fc8c0bca5d471cc9.
