+++ epitran/test/test_yue.py
@@ -8,6 +8,8 @@ import epitran
 
 
 class TestYue(unittest.TestCase):
+    
+
     def setUp(self):
         self.epi = epitran.Epitran('yue-Latn')
         
@@ -27,10 +29,40 @@ class TestYue(unittest.TestCase):
             ('maa5','ma˨˧') #马，妈
         ]
 
+
     def test(self):
+        
         for ex in self.tester_exs: 
             jyuping, ipa = ex
             pred = self.epi.transliterate(jyuping)
             self.assertEqual(pred, ipa)
             # if pred != ipa: 
             #     print(f'{pred} - {ipa}')
+
+class TestYue2(unittest.TestCase):
+
+    def setUp(self):
+        self.epi = epitran.Epitran('yue-Hant', tones=True)
+        
+        # wikipedia derived test cases 
+        self.tester_exs = [
+            ('心','sɐm˥'),
+            ('手','sɐw˨˥'),
+            ('一','jɐt̚˥'),
+            ('有','jɐw˨'),
+            ('猪','tsy˥'),
+            ('话','wa˨˥'),
+            ('人','jɐn˨˩'),
+            ('鋸','kœ˥'),
+            ('天','tʰin˥'),
+            ('好','hɔw˧'),
+            ('都','tɔw˥'),
+            ('媽','ma˨˧'),
+        ]
+
+    def test(self):
+        
+        for ex in self.tester_exs: 
+            jyuping, ipa = ex
+            pred = self.epi.transliterate(jyuping)
+            self.assertEqual(pred, ipa)
